pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lab1-serialcomm"
version = "1.0.0"
description = "Simple serial communication GUI for Arduino/ESP32 sensors"
authors = [{ name = "Pranshul Shenoy", email = "pranshul.shenoy@research.iiit.ac.in" }, {name = "Neha Prabhu", email = "neha.prabhu@students.iiit.ac.in"}, {name = "Samarth Rao", email = "samarth.rao@research.iiit.ac.in"}, {name = "Avani Sood", email = "avani.sood@research.iiit.ac.in"}]
readme = "README.md"
requires-python = ">=3.9"

# Core dependencies - keep it minimal
dependencies = [
    "pyserial>=3.5",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "tomli>=2.0.0; python_version<'3.11'",
]

# Optional dependencies
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
]

# Simple entry point (console script)
[project.scripts]
lab1-serialcomm = "serial_monitor.main:main"

# Hatch/hatchling build configuration
# Use the package *path(s)* to point hatchling at your src/ layout.
[tool.hatch.build.targets.wheel]
packages = ["src/serial_monitor"]

# (Optional) sdist target: include only the package directory (keeps sdist small)
[tool.hatch.build.targets.sdist]
only-include = ["src/serial_monitor"]

# Tool configs
[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
fix = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v", "--tb=short"]
